requirements.txt
streamlit
pillow
imageio
imageio-ffmpeg
